Fixed ignorance of arguments of expression
And cleaned/noted FilterMatcher.tvesMatch() function.
